refactor Unit test
